libtomcrypt: implement zeromem() with memzero_explicit()
This change is part of commit 03121b2ca443 ("core: crypto: libtomcrypt:
fix LTC_CLEAN_STACK bug"). LTC_CLEAN_STACK will be disabled in a later
commit because it is not supported anymore upstream ("LTC_CLEAN_STACK
is considered as broken"), so we keep only the zeromem() part of the
commit.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>