ta: avb: solve an issue to write the AVB lock state in secure storage
When writing the lock state, write_lock_state() returns
TEE_ERROR_ACCESS_CONFLICT. This error occurs because
TEE_CreatePersistentObject() tries to overwrite the file
but the file was already opened by TEE_OpenPersistentObject().
Use TEE_SeekObjectData() and TEE_WriteObjectData() instead.

Note that, when the lock state is updated, the rollback
indexes are preserved.

Signed-off-by: Alexandre Jutras <alexandre.jutras@linaro.org>