ftrace: use wall clock time
The current ftrace timestamping code for TAs ignores the time spent
outside user space. While this has positive aspects, such as avoiding
noise due to context switching, it can be misleading. I also suspect it
will not work well with combined kernel + user space tracing (when
syscall ftrace is implemented).

This commit removes the code that was specially inserted to ignore
non-user time. As a result, the durations shown in the dump are computed
from the wall clock times taken on function entries and exits.
In my opinion, this gives a much better view of the actual execution.
Here is an example with xtest 1009 (only the part that calls
TEE_Wait(2000)), built for QEMU with:

 $ make CFG_TA_FTRACE_SUPPORT=y CFG_ULIBS_MCOUNT=y \
        CFLAGS_ta_arm32="-O0 -pg"

 $ cat ftrace-5b9e0e40-2636-11e1-ad9e0002a5d5c51b.out | \
   ./optee_os/scripts/symbolize.py -d out-br/build/optee_test-1.0/ta/*/out

 Function graph for TA: 5b9e0e40-2636-11e1-ad9e-0002a5d5c51b @ 170000
 ...
              | __ta_entry_c() {
              |  __utee_entry() {
              |   entry_invoke_command() {
    21.120 us |    ta_header_get_session();
    25.040 us |    from_utee_params();
              |    ta_header_save_params() {
     1.072 us |     memcpy();
    13.728 us |    }
              |    TA_InvokeCommandEntryPoint() {
              |     ta_entry_wait() {
              |      printf() {
    11.888 us |       trace_get_level();
              |       vsnprintf() {
              |        __vsnprintf() {
   387.984 us |         kprintf();
   451.840 us |        }
   508.128 us |       }
              |       trace_ext_puts() {
    56.192 us |        strlen();
   371.440 us |        utee_log();
   528.960 us |       }
  1143.376 us |      }
              |      TEE_Wait() {
  2026.760 ms |       utee_wait();
  2027.216 ms |      }
  2028.504 ms |     }
  2028.646 ms |    }
   257.312 us |    to_utee_params();
  2029.316 ms |   }
              |   ta_header_save_params() {
     9.216 us |    memset();
    94.528 us |   }
  2029.599 ms |  }
  2029.670 ms | }
 ...

Fixes: f5df167c2ffb ("ftrace: Add function execution time support")
Signed-off-by: Jerome Forissier <jerome@forissier.org>