hikey: use default (64 kB) core heap size when pager is on
Commit 28c75dbebc49 ("hikey: increase core heap size to 192 kB") has
increased the core heap size in order to pass the AOSP VTS.
Unfortunately, this bigger value does not work well when the pager is
enabled: it causes lots of page faults and a massive slowdown (for
instance, 'xtest 1013' on HiKey620 completes in ~ 1.7 s with the
default heap size of 64 kB but takes ~ 53 s with 192 kB).

Since the configuration used for AOSP VTS does not include the pager,
let's stick to the default heap size in this case.

Signed-off-by: Jerome Forissier <jerome@forissier.org>
Acked-by: Victor Chong <victor.chong@linaro.org>