core: no boot thread for 32bit mode when CFG_VIRTUALIZATION=y
This change fixes 32bit mode configuration when CFG_VIRTUALIZATION=y
in which case the boot thread is not used hence must not be cleared,
as done in 64bit mode configuration.

This change also modifies the 64bit implementation to include the
inline comment into conditioned instruction sequence to highlight
clearing boot thread must not be done when CFG_VIRTUALIZATION=y.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>