core: generic_boot: extend dts while supporting overlay generation.
The following commit will insert the OP-TEE node in a device tree
passed as an input parameter during boot while it will also generate
an OP-TEE specific overlay at a build specific address (CFG_OVERLAY_ADDR).

This functionality is useful when it is SPL that boots OP-TEE which in
turns boots U-boot; then U-Boot will contain the correct OP-TEE node
plus an overlay which U-boot can then apply to the linux dtb before
starting it.

When this feature is enabled, CFG_DT_ADDR and CFG_EXTERNAL_DTB_OVERLAY
are not required.

Signed-off-by: Jorge Ramirez-Ortiz <jorge@foundries.io>