drivers: imx_wdog: set functional state prior restart.
The linux kernel is responsible for probing and initializing the
imx watchdog. It is therefore possible for any subsystem -ie
systemd- to stop and disable the watchdog as part of the reboot
house-keeping tasks.
If this happens and the kernel calls OP-TEE's psci to reboot, the
watchdog will be in the wrong state to perform the requested reset and
the system will hang in a zombie state.

This commit re-initializes the watchdog to a ready state before
requesting reset.

This use case has been tested on imx7ulp.

Signed-off-by: Jorge Ramirez-Ortiz <jorge@foundries.io>