[Review] Add libsm implementing the Chinese SM2/SM3/SM4 algorithms for TAs
I will append the following to the initial commit description:

"
The publication date of the GP Internal Core API v1.3 specification is
not known, and we don't know what the SM APIs will look like at this
point. So, rather than trying to make guesses, which could cause hard
tp fix incompatibilities, it is more sane to create a separate, small
user space library as proposed in this commit. When the GP API is
finally implemented, we can decide whether or not to deprecate the
library or continue maintaining it (it is not that big in anyways).
"

Signed-off-by: Jerome Forissier <jerome@forissier.org>