core: crypto: add support for SM2 PKE
SM2 is a set of public key cryptographic algorithms based on elliptic
curves published by the Chinese Commercial Cryptography Administration
Office [1]. The standard has three main parts which define:

 - A Digital Signature Algorithm (part 2)
 - A Key Exchange Protocol (part 3)
 - A Public Key Encryption method (part 4)

This commit implements the PKE. As far as the ECC point encoding goes,
only the uncompressed form is supported, as permitted by the standard.
In practice, it means that the cipher text always starts with the byte
value 0x04. Other forms cannot be decrypted by this implementation.

SM2 is enabled with CFG_CRYPTO_SM2_PKE=y which currently requires that
CFG_CRYPTOLIB_NAME=tomcrypt. An Mbed TLS implementation could be added
later if needed.

Signed-off-by: Jerome Forissier <jerome@forissier.org>