plat-stm32mp1: define memory dma to shared resource
This change defines memory DMA as a shared resource. Despite
Secure and non-secure have have specific access to memory
DMA, the reset control is exclusive to the secure world.

With memory DMAs defined as the shared resource, the secure
world will be able to open access to the resource if it is
not used by the secure side.

Signed-off-by: Etienne Carriere <etienne.carriere@st.com>