plat-stm32mp1: scmi server reset controllers in non-secure world
Embed a SCMI server in stm32mp1 to handle SCMI reset domain requests
from the non-secure world for resource that, because of secure
hardening of the system, are restricted to secure world accesses only.

The communication channel for SCMI message exchange is a small shared
memory buffer is SYSRAM, manage with a SMT header. CPUs enters
SCMI message drivers through a fastcall SMC related to a platform
SiP SMC function ID.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>