libutils: add strtok
The strtok() function breaks a string into a sequence of zero or more
nonempty tokens.  On the first call to strtok(), the string to be
parsed should be specified in str.  In each subsequent call that
should parse the same string, str must be NULL.

Each call to strtok() returns a pointer to a null-terminated string
containing the next token.  This string does not include the
delimiting byte.

If no more tokens are found, strtok() returns NULL.

Signed-off-by: Jorge Ramirez <jorge@foundries.io>